[refactor]: change folder structure (#8)
