+++ src/components/commons/index.ts
@@ -1,5 +1,5 @@
 // Common UI Components
-export { default as TopBar } from "./TopBar";
+export { default as TopBar } from "./TopBar"
 export { default as Select } from "./Select";
 export { default as SearchInput } from "./SearchInput";
 export { default as ViewToggle } from "./ViewToggle";
