+++ src/components/commons/sidebar/MobileSidebar.tsx
@@ -1,8 +1,8 @@
 ﻿import React, { useState } from "react";
 
 import menuData from "@/constants/SidebarConstants";
-import SideTitleList from "@/commons/sidebar/SideTitleList";
-import SideDetailList from "@/commons/sidebar/SideDetailList";
+import SideTitleList from "@/components/commons/sidebar/SideTitleList";
+import SideDetailList from "@/components/commons/sidebar/SideDetailList";
 
 const MobileSidebar: React.FC = () => {
   // const { closeSidebar } = useSidebarStore();
