+++ src/components/commons/sidebar/SideDetailItem.tsx
